refactor: css 관심사 분리하기
