+++ src/App.css.ts
@@ -1,4 +1,4 @@
-import { globalStyle, style } from '@vanilla-extract/css';
+import { globalStyle } from '@vanilla-extract/css';
 
 globalStyle('.title1', {
   fontWeight: 700,
@@ -76,42 +76,4 @@ globalStyle('.text-secondary', {
 
 globalStyle('.text-subtitle', {
   color: 'var(--text-subtitle)'
-});
-
-export const wrapper = style({
-  paddingInline: 24
-});
-
-export const main = style({
-  margin: '0 auto',
-  maxWidth: 'var(--max-width)'
-});
-
-export const header = style({
-  display: 'flex',
-  justifyContent: 'center',
-  alignItems: 'center',
-  position: 'relative',
-  height: 80
-});
-
-export const heading = style({
-  position: 'absolute',
-  left: 0,
-  textTransform: 'uppercase'
-});
-
-export const nav = style([main]);
-
-export const list = style({
-  display: 'flex',
-  gap: 56
-});
-
-export const item = style({});
-
-export const linkActive = style({
-  textDecoration: 'underline',
-  textDecorationColor: 'var(--palette-primary)',
-  textUnderlineOffset: 8
 });
